detection.c
Our project mainly focuses on Human Detection based on information stored as a reference, which is locked as a target. Then the obtained information from the real time scenario is also stored. Both the conditions are being compared and then the difference is then analyzed. Through this project movement and detection can be analyzed. Basic movements like jumping,crawling...etc produces a beep sound which is also a prominent feature of this software.